+++ tests/protocols/gmpv208/test_v7_from_gmpv208.py
@@ -266,10 +266,6 @@ class Gmpv208GetCredentialsTestCase(GmpGetCredentialsTestCase, Gmpv208TestCase):
     pass
 
 
-class Gmpv208GetFeedTestCase(GmpGetFeedTestCase, Gmpv208TestCase):
-    pass
-
-
 class Gmpv208GetFeedsTestCase(GmpGetFeedsTestCase, Gmpv208TestCase):
     pass